feat: 添加键盘交互事件.
Description: 添加了键盘操作函数.

Log: 可通过上下左右键操作.

link:

+++ src/widgets/borderwidget.cpp
@@ -42,7 +42,7 @@ void BorderWidget::setChecked(bool checked)
 
 void BorderWidget::updateInterface(float f)
 {
-    m_size = m_pixmap.size() * f;
+    m_size = m_pixmap.size() * static_cast<qreal>(f);
     update();
 }
